// src/app/components/moduleusers/models/Sucursales.ts
export class Sucursales{
    idsucursal?: number;
    codigosucursal?: string;
    nombresucursal?: string;
    descripcionsucursal?: string;
    direccionsucursal?: string;
    telefonosucursal?: string;
    status?: boolean;
    fechacreacion?: string;
    fechamodificacion?: string;
    ipcreacion?: string;
    ipmodificacion?: string;
    fecharegistro?: string;
    idpais?: number;
    idcorporacion?: number;
    url?: string;


}