// src/app/components/moduleusers/models/Departamentos.ts
export class Departamentos{
    iddepartamento?: number;
    codigodepartamento?: string;
    nombredepartamento?: string;
    status?: boolean;
    descripciondepartamento?: string;
    fechacreacion?: string;
    fechamodificacion?: string;
    ipcreacion?: string;
    ipmodificacion?: string;
    fecharegistro?: string;
    idsucursal?: number;
    url?: string;


}